requirements.txt
python-telegram-bot==8.0
SQLAlchemy==1.1.14
feedparser==5.2.1
